+++ bin/irregular/rebuild_duckdb.dart
@@ -1,6 +1,5 @@
 import 'package:date/date.dart';
 import 'package:elec/elec.dart';
-import 'package:elec_server/src/db/lib_prod_archives.dart';
 import 'package:elec_server/src/db/lib_update_dbs.dart';
 import 'package:logging/logging.dart';
 import 'package:timezone/data/latest.dart';